+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ Changelog
 https://gitlab.com/gf-bl/oereb_client/milestones/13
 
 - Build and upload tagged versions to public PyPI
+- Set title and favicon via configuration
 
 1.1.5
 *****